different urls for different topcis

+++ src/utils.js
@@ -11,37 +11,44 @@ export function fetchArticles() {
 }
 
 export function fetchArticleById(id) {
-  return ncNewsApi.get(`/api/articles/${id}`).then((res) => {
-    return res.data;
-  });
+  if (Number(id)) {
+    return ncNewsApi.get(`/api/articles/${id}`).then((res) => {
+      return res.data;
+    });
+  }
 }
 
 export function fetchCommentsByArticleId(id) {
-  return ncNewsApi.get(`/api/articles/${id}/comments`).then((res) => {
-    return res.data.commentsArray;
-  });
+  if (Number(id)) {
+    return ncNewsApi.get(`/api/articles/${id}/comments`).then((res) => {
+      return res.data.commentsArray;
+    });
+  }
 }
 
 export function patchArticleById(id, vote) {
-  return ncNewsApi.patch(`/api/articles/${id}`, { inc_votes: vote });
+  if (Number(id)) {
+    return ncNewsApi.patch(`/api/articles/${id}`, { inc_votes: vote });
+  }
 }
 
 export function postCommentByArticleId(id, comment) {
-  return ncNewsApi
-    .post(`/api/articles/${id}/comments`, {
-      username: comment.username,
-      body: comment.body,
-    })
-    .then((res) => res.data.commentObj);
+  if (Number(id)) {
+    return ncNewsApi
+      .post(`/api/articles/${id}/comments`, {
+        username: comment.username,
+        body: comment.body,
+      })
+      .then((res) => res.data.commentObj);
+  }
 }
 
 export function deleteCommentById(id) {
-  return ncNewsApi.delete(`/api/comments/${id}`)
+  if (Number(id)) return ncNewsApi.delete(`/api/comments/${id}`);
 }
 
 export function getTopics() {
-  return ncNewsApi.get(`/api/topics`)
-  .then((res) => {
-    return res.data
-  })
+  return ncNewsApi.get(`/api/topics`).then((res) => {
+    return res.data;
+  });
 }
